EventMate/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Agent framework
uagents==0.12.1

# AI/ML libraries
openai==1.3.8
groq==0.4.1

# Data handling
pydantic==2.5.1

# Environment management
python-dotenv==1.0.0

# HTTP requests
httpx==0.25.2

# CORS middleware (included in FastAPI but listed for clarity)
# fastapi already includes starlette which provides CORS

# Type hints
typing-extensions==4.8.0

# Additional utilities
asyncio-mqtt==0.13.0
